+++ bricks/ev3/mpconfigport.h
@@ -42,6 +42,7 @@
 #define PYBRICKS_PY_IODEVICES_UART_DEVICE       (1)
 #define PYBRICKS_PY_IODEVICES_XBOX_CONTROLLER   (0)
 #define PYBRICKS_PY_MESSAGING                   (1)
+#define PYBRICKS_PY_MESSAGING_APP_DATA          (1)
 #define PYBRICKS_PY_MESSAGING_RFCOMM            (1)
 #define PYBRICKS_PY_MESSAGING_BLE_RADIO         (0)
 #define PYBRICKS_PY_NXTDEVICES                  (1)
@@ -58,7 +59,6 @@
 #define PYBRICKS_PY_ROBOTICS_DRIVEBASE_SPIKE    (0)
 #define PYBRICKS_PY_TOOLS                       (1)
 #define PYBRICKS_PY_TOOLS_HUB_MENU              (0)
-#define PYBRICKS_PY_TOOLS_APP_DATA              (1)
 
 // Pybricks options
 #define PYBRICKS_OPT_COMPILER                   (1)