+++ bricks/virtualhub/mpconfigport.h
@@ -39,6 +39,7 @@
 #define PYBRICKS_PY_IODEVICES_UART_DEVICE       (0)
 #define PYBRICKS_PY_IODEVICES_XBOX_CONTROLLER   (1)
 #define PYBRICKS_PY_MESSAGING                   (1)
+#define PYBRICKS_PY_MESSAGING_APP_DATA          (1)
 #define PYBRICKS_PY_MESSAGING_RFCOMM            (1)
 #define PYBRICKS_PY_MESSAGING_BLE_RADIO         (1)
 #define PYBRICKS_PY_MESSAGING_BLE_RADIO_OLD     (1)
@@ -58,7 +59,6 @@
 #define PYBRICKS_PY_ROBOTICS_DRIVEBASE_SPIKE    (0)
 #define PYBRICKS_PY_TOOLS                       (1)
 #define PYBRICKS_PY_TOOLS_HUB_MENU              (0)
-#define PYBRICKS_PY_TOOLS_APP_DATA              (1)
 
 // Pybricks options
 #define PYBRICKS_OPT_COMPILER                   (1)