+++ pybricks/hubs/pb_type_primehub.c
@@ -23,8 +23,9 @@
 #include <pybricks/util_pb/pb_error.h>
 
 #include <pybricks/common.h>
-#include <pybricks/tools/pb_type_matrix.h>
 #include <pybricks/hubs.h>
+#include <pybricks/messaging/messaging.h>
+#include <pybricks/tools/pb_type_matrix.h>
 
 typedef struct _hubs_PrimeHub_obj_t {
     mp_obj_base_t base;